+++ inst/scripts/demo.R
@@ -1,6 +1,7 @@
 library(sager)
 library(Spectra)
 library(QFeatures)
+library(SummarizedExperiment)
 
 ## ---------------------------
 ## With Spectra
@@ -10,6 +11,10 @@ sp$filename <- basename(dataOrigin(sp))
 sp <- makeKEY(sp, var = c("filename", "spectrumId"))
 sp$.KEY
 
+
+sp <- makeKEY(sp, var = c("dataOrigin", "spectrumId"), key = "KEY2")
+
+subsetByKEY(sp, "5779224eb8e7_sager_subset_PXD016766.mzML.controllerType=0 controllerNumber=1 scan=5869")
 
 qf <- sageQFeatures(sagerQuantData(), sagerIdData())
 
@@ -22,6 +27,6 @@ rowData(se)$.KEY
 ## With QFreature
 
 qf <- makeKEY(qf, vars = c("file", "scannr"))
-rowData(qf[[1]])$.KEY
-rowData(qf[[2]])$.KEY
+head(rowData(qf[[1]])$.KEY)
+head(rowData(qf[[2]])$.KEY)
 rowData(qf[[3]])$.KEY